requirements.txt
streamlit
pandas
plotly
openpyxl

crisp-api
python-dotenv

textblob
